Extract transformation cage to a seperate file
