+++ editor/src/viewport_tools/tools/shared/mod.rs
@@ -1,1 +1,2 @@
 pub mod resize;
+pub mod transformation_cage;
